refactor: move team partition scheme to its own module
